@@ -18,27 +18,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.7",
-    "@angular/common": "7.2.7",
-    "@angular/compiler": "7.2.7",
-    "@angular/core": "7.2.7",
-    "@angular/fire": "5.1.1",
-    "@angular/forms": "7.2.7",
+    "@angular/animations": "8.2.0",
+    "@angular/common": "8.2.0",
+    "@angular/compiler": "8.2.0",
+    "@angular/core": "8.2.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "8.2.0",
     "@angular/http": "7.2.7",
-    "@angular/platform-browser": "7.2.7",
-    "@angular/platform-browser-dynamic": "7.2.7",
-    "@angular/router": "7.2.7",
+    "@angular/platform-browser": "8.2.0",
+    "@angular/platform-browser-dynamic": "8.2.0",
+    "@angular/router": "8.2.0",
     "core-js": "^2.6.5",
     "firebase": "^5.8.4",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "tslib": "^1.9.3",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
     "@angular/cli": "7.3.3",
-    "@angular/compiler-cli": "7.2.7",
-    "@angular/language-service": "7.2.7",
+    "@angular/compiler-cli": "8.2.0",
+    "@angular/language-service": "8.2.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^6.14.3",
@@ -54,6 +54,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~3.2.0",
     "tslint": "^5.13.0",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   }
 }
